requirements.txt
aniso8601==8.1.1
api==0.0.7
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
conda==4.3.16
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
nose==1.3.7
pipenv==2020.8.13
psycopg2-binary==2.8.6
pycosat==0.6.3
pytz==2021.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
Werkzeug==1.0.1
